LG-1410: rename gem to identity_validations
