--- lib/identity/validations.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-require 'identity/validations/version'
-require 'identity/validations/service_provider_validation'
-
-module Identity
-  module Validations
-    class Error < StandardError; end
-    # Your code goes here...
-  end
-end
