--- lib/identity/validations/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Identity
-  module Validations
-    VERSION = '0.1.0'
-  end
-end
