+++ lib/identity_validations.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'identity_validations/version'
+require 'identity_validations/service_provider_validation'
+
+module IdentityValidations
+  class Error < StandardError; end
+  # Your code goes here...
+end
