+++ lib/identity_validations/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module IdentityValidations
+  VERSION = '0.1.0'
+end
