+++ spec/identity/validations_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-RSpec.describe Identity::Validations do
+RSpec.describe IdentityValidations do
   it 'has a version number' do
-    expect(Identity::Validations::VERSION).not_to be nil
+    expect(IdentityValidations::VERSION).not_to be nil
   end
 end
