+++ spec/spec_helper.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'bundler/setup'
-require 'identity/validations'
+require 'identity_validations'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
